Split out AuthorizationHeader to its own file
This commit moves the `AuthorizationHeader` enum into its own file to
help keep mod.rs clean.
